Put yamlMetaBlock parser into separate module
The YAML metadata parser is useful for many formats and is specific
enough to warrant a separate module.
